+++ src/javascript/types/puffin.component.ts
@@ -1,0 +1,7 @@
+export interface PuffinComponent {
+	_type: string
+	_props: any
+	_opened: boolean
+	_is: string
+	children: this[]
+}
