+++ webpack.config.js
@@ -14,7 +14,7 @@ module.exports = {
 	plugins: [
 		new WebpackMessages({
 			name: 'Graviton',
-			logger: str => console.log(`--> ${str}`),
+			logger: str => console.log(`[webpack] --> ${str}`),
 		}),
 		new HtmlWebpackPlugin({
 			title: 'Graviton',
@@ -84,6 +84,9 @@ module.exports = {
 			AppPlatform: path.resolve(__dirname, './src/javascript/utils/platform.ts'),
 			Constants: path.resolve(__dirname, './src/javascript/defaults/constants.ts'),
 			StaticPlugins: path.resolve(__dirname, './plugins/static'),
+			Types: path.resolve(__dirname, './src/javascript/types'),
+			Constructors: path.resolve(__dirname, './src/javascript/constructors'),
+			Components: path.resolve(__dirname, './src/javascript/components'),
 		},
 	},
 	target: 'electron-renderer',
